--- frontend/src/components/inventory-form_addItem.tsx
@@ -1,122 +1,0 @@
-"use client";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-
-import {
-Form,
-FormControl,
-FormField,
-FormItem,
-FormLabel,
-FormMessage,
-} from "@/components/ui/form";
-
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-
-const formSchema = z.object({
-    item: z
-        .string()
-        .min(2, {
-            message: "Brand name must contain 2 - 50 characters.",
-        })
-        .max(50),
-    quantity: z
-        .number({ required_error: "Please enter a quantity" })
-        .min(1, { message: "Quantity must be at least 1." }),
-    price: z
-        .number({ required_error: "Please enter a price" })
-        .min(0.01, { message: "Price must be at least 0.01." }),
-    expirationDate: z
-        .date({ required_error: "Please enter an expiration date" })
-        .refine((date) => date > new Date(), {
-            message: "Expiration date must be in the future.",
-        }),
-});
-
-export default function ProfileForm() {
-// 1. Define your form.
-const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-        item: "",
-        quantity: 1,
-        price: 0.01,
-        expirationDate: new Date(),
-    },
-});
-
-// 2. Define a submit handler.
-function onSubmit(values: z.infer<typeof formSchema>) {
-    // Do something with the form values.
-    // ✅ This will be type-safe and validated.
-    console.log(values);
-}
-
-return (
-    <Form {...form}>
-        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8 w-full">
-            <FormField
-                control={form.control}
-                name="item"
-                render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Item Name</FormLabel>
-                        <FormControl>
-                            <Input placeholder="Enter item name here" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-            />
-            <FormField
-                control={form.control}
-                name="quantity"
-                render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Quantity</FormLabel>
-                        <FormControl>
-                            <Input type="number" placeholder="Enter quantity here" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-            />
-            <FormField
-                control={form.control}
-                name="price"
-                render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Price</FormLabel>
-                        <FormControl>
-                            <Input type="number" step="0.01" placeholder="Enter price here" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-            />
-            <FormField
-                control={form.control}
-                name="expirationDate"
-                render={({ field }) => (
-                    <FormItem>
-                        <FormLabel>Expiration Date</FormLabel>
-                        <FormControl>
-                            <Input
-                                type="date"
-                                {...field}
-                                value={field.value ? field.value.toISOString().split('T')[0] : ''}
-                                onChange={(e) => field.onChange(e.target.value ? new Date(e.target.value) : null)}
-                            />
-                        </FormControl>
-                        <FormMessage />
-                    </FormItem>
-                )}
-            />
-        </form>
-        <Button type="submit">Submit</Button>
-    </Form>
-);
-}
